refactor(qa_agent): reorganiza estrutura de pastas seguindo padrão dos demais agentes
Divide o monolito receive_requirements.py (677 linhas) em pacote com
responsabilidades isoladas (agent, prompt, io, normalizer, sanitizer,
llm_generation, orchestration); extrai a instruction inline do
code_fix_agent para prompt.py; renomeia qa_prompt.py para prompt.py
(padrão dos outros agentes); remove doubt_tool.py, duplicata morta e
sem uso do DoubtArtifactGenerator já existente em shared/tools/.

Mudança puramente estrutural — nenhuma lógica ou comportamento alterado
(suíte de testes com os mesmos resultados antes/depois).
