+++ adk/src/agents/qa_agent/subagents/code_fix_agent/agent.py
@@ -1,17 +1,14 @@
 import os
 from google.adk.agents import LlmAgent
-from google.adk.agents import Agent
 from google.adk.tools import FunctionTool
 from shared.tools.build_fix_prompt import build_fix_prompt_from_error, build_fix_prompt_from_pytest
+
+from .prompt import CODE_FIX_AGENT_PROMPT
 
 agent = LlmAgent(
     name="code_fix_agent",
     model=os.environ.get("ADK_LLM_MODEL", "gemini-2.5-flash"),
-    instruction=(
-                "Você é um subagente de um agente de teste"
-                "Seu papel é receber logs de funcionamento de programas, entender o que ele significa, e, se ver algum erro, fazer um prompt de correção para um outro agente que irá realizar a codificação."
-                "Para isso, você tem as ferramentas de parsing de logs, que irão resumir o log para você, e as ferramentas de construir o prompt"
-                ),
+    instruction=CODE_FIX_AGENT_PROMPT,
     description=(
         "Recebe logs e relatórios de falha de testes (pytest) e gera prompts de correção "
         "focados na causa raiz, alimentando o ciclo de autocorrect do agente QA."
