--- adk/src/agents/qa_agent/subagents/receive_requirements.py
@@ -1,677 +1,0 @@
-import ast
-import asyncio
-import base64
-import json
-import logging
-import os
-import re
-from datetime import datetime, timezone
-from pathlib import Path
-
-from google.adk.agents import LlmAgent
-from google.adk.tools import FunctionTool
-import litellm
-from litellm import completion
-
-# Garante compatibilidade com providers que não suportam response_format
-# (ex.: github_copilot). Precisa estar antes de qualquer chamada a completion().
-litellm.drop_params = True
-
-from shared.workspace import get_agent_workspace
-
-logger = logging.getLogger("qa_agent")
-
-
-def _tests_dir() -> Path:
-    """workspace_output/tests/inputs/ resolvido em runtime.
-
-    Centraliza o destino dos arquivos pytest gerados pelo subagente.
-    Resolvido em runtime para respeitar WORKSPACE_OUTPUT_DIR env var.
-    """
-    return get_agent_workspace("receive_requirements")
-
-
-def _doubt_dir() -> Path:
-    """Sibling 'doubt_artifacts' dentro do diretório de testes."""
-    return _tests_dir() / "doubt_artifacts"
-
-# ──────────────────────────────────────────────────────────────────────────────
-# Ponto de entrada público (registrado como FunctionTool no agent.py)
-# ──────────────────────────────────────────────────────────────────────────────
-
-def _run_async(coro):
-    """Executa coroutine de forma segura com ou sem event loop ativo.
-
-    Args:
-        coro: Coroutine a ser executada.
-
-    Returns:
-        Resultado da coroutine.
-
-    Note:
-        Compatível com FastAPI + ADK - usa ThreadPoolExecutor se houver loop ativo.
-    """
-    try:
-        asyncio.get_running_loop()
-        # Há um loop rodando (FastAPI/ADK) — executa em thread separada
-        import concurrent.futures
-        with concurrent.futures.ThreadPoolExecutor() as pool:
-            future = pool.submit(asyncio.run, coro)
-            return future.result()
-    except RuntimeError:
-        # Sem loop rodando — pode usar asyncio.run normalmente
-        return asyncio.run(coro)
-
-def _parse_fragmented_requirements(raw_input: str) -> list:
-    """Converte texto livre/fragmentado em uma lista estruturada de artefatos."""
-    prompt = f"""Extraia os requisitos do texto abaixo e retorne um JSON array estrito no formato:
-[{{ "id_artefato": "RF-001", "tipo": "RF", "conteudo": "...", "modulo": "...", "criticidade": "alta|media|baixa" }}]
-Identifique os tipos (RF, RNF, HU, UC, RN). Se não houver ID claro, gere um sequencial.
-Texto bruto: {raw_input}
-"""
-    model_name = os.environ.get("ADK_LLM_MODEL", "gemini-2.5-flash")
-    llm_kwargs = {}
-    if "/" not in model_name:
-        model_name = f"gemini/{model_name}"
-        llm_kwargs["api_key"] = os.environ.get("GOOGLE_API_KEY")
-    response = completion(
-        model=model_name,
-        messages=[{"role": "user", "content": prompt}],
-        temperature=0,
-        **llm_kwargs,
-    )
-    
-    conteudo = response.choices[0].message.content.strip()
-    # Limpa possíveis formatações markdown do retorno do LLM
-    if conteudo.startswith("```json"):
-        conteudo = conteudo.replace("```json\n", "").replace("```", "")
-    return json.loads(conteudo)
-
-def _extrair_de_parts(parts: list) -> tuple[list[dict], list[str]]:
-    arquivos_apoio: list[dict] = []
-    textos: list[str] = []
-
-    for part in parts:
-        if not isinstance(part, dict):
-            continue
-
-        texto = part.get("text")
-        if isinstance(texto, str) and texto.strip():
-            textos.append(texto.strip())
-
-        inline = part.get("inlineData")
-        if not isinstance(inline, dict):
-            continue
-
-        nome = inline.get("displayName") or inline.get("name") or "arquivo_anexo"
-        mime = inline.get("mimeType") or "application/octet-stream"
-        conteudo_b64 = inline.get("data")
-        if not isinstance(conteudo_b64, str) or not conteudo_b64.strip():
-            continue
-
-        entrada_arquivo = {"nome": nome}
-        try:
-            bruto = base64.b64decode(conteudo_b64)
-            texto_decodificado = bruto.decode("utf-8")
-            entrada_arquivo["conteudo"] = texto_decodificado
-            if texto_decodificado.strip():
-                textos.append(texto_decodificado.strip())
-        except Exception:
-            entrada_arquivo["conteudo_base64"] = conteudo_b64
-
-        entrada_arquivo["mime_type"] = mime
-        arquivos_apoio.append(entrada_arquivo)
-
-    return arquivos_apoio, textos
-
-def _normalizar_anexos_inline(lista_artefatos: list) -> list:
-    normalizados = []
-    for artefato in lista_artefatos:
-        if not isinstance(artefato, dict):
-            continue
-
-        arquivos_existentes = artefato.get("arquivos_apoio", [])
-        if not isinstance(arquivos_existentes, list):
-            arquivos_existentes = []
-
-        parts = artefato.get("parts")
-        if not isinstance(parts, list):
-            content = artefato.get("content", {})
-            if isinstance(content, dict):
-                parts = content.get("parts")
-
-        textos_extraidos: list[str] = []
-        if isinstance(parts, list):
-            arquivos_extraidos, textos_extraidos = _extrair_de_parts(parts)
-            arquivos_existentes.extend(arquivos_extraidos)
-
-        if arquivos_existentes:
-            artefato["arquivos_apoio"] = arquivos_existentes
-
-        conteudo_atual = artefato.get("conteudo", "")
-        if not isinstance(conteudo_atual, str):
-            conteudo_atual = str(conteudo_atual)
-
-        if textos_extraidos:
-            bloco_extra = "\n\n".join(textos_extraidos).strip()
-            if bloco_extra and bloco_extra not in conteudo_atual:
-                artefato["conteudo"] = (
-                    f"{conteudo_atual}\n\n{bloco_extra}".strip()
-                    if conteudo_atual.strip()
-                    else bloco_extra
-                )
-
-        normalizados.append(artefato)
-
-    return normalizados
-
-
-def _validar_e_sanitizar_codigo(codigo: str, id_artefato: str) -> str:
-    """Sanitiza tokens fora-da-gramática Python e valida via ast.parse.
-
-    Aplica regex que remove placeholders entre `<>` colocados após keywords
-    Python (pass<X>, return<Y>, etc.) e em seguida valida o código com
-    ast.parse. Se mesmo após sanitização o código permanece inválido,
-    levanta ValueError — o chamador propaga o erro para o autocorrect cycle.
-
-    Args:
-        codigo: String com código Python emitido pelo LLM.
-        id_artefato: ID do artefato (usado nas mensagens de log/erro).
-
-    Returns:
-        Código Python sanitizado e validado.
-
-    Raises:
-        ValueError: Se ast.parse falha após sanitização.
-    """
-    padrao = re.compile(r'\b(pass|return|continue|break|raise)<[^>\n]*>')
-    sanitizado = padrao.sub(r'\1', codigo)
-
-    if sanitizado != codigo:
-        logger.warning(
-            f"[QA] Sanitização aplicada em {id_artefato}: "
-            f"removidos placeholders fora da gramática Python."
-        )
-
-    try:
-        ast.parse(sanitizado)
-    except SyntaxError as e:
-        raise ValueError(
-            f"Código gerado para {id_artefato} é inválido após sanitização: "
-            f"{e.msg} (linha {e.lineno}). Será reciclado via autocorrect."
-        ) from e
-
-    return sanitizado
-
-
-def receber_requisitos(artefatos_json: str) -> dict:
-    """
-    Recebe uma lista de artefatos de requisito em JSON e gera testes pytest
-    em paralelo para cada um.
-
-    Args:
-        artefatos_json: String JSON com um ou mais artefatos.
-            Formato de um artefato:
-            {
-                "id_artefato": "RF-001",
-                "tipo": "RF",
-                "conteudo": "O sistema deve...",
-                "modulo": "nome_do_modulo",
-                "criticidade": "alta"
-            }
-
-    Returns:
-        Relatório consolidado:
-        {
-            "status": "concluido",
-            "resumo": { "total": N, "sucessos": N, "bloqueados": N, "falhas": N },
-            "detalhes": [ ... ]
-        }
-    """
-    try:
-        lista = json.loads(artefatos_json)
-        if isinstance(lista, dict):
-            lista = [lista]
-    except json.JSONDecodeError as e:
-        logger.warning(f"[QA] Falha ao ler JSON estrito. Tentando extrair de fragmentos de texto...")
-        try:
-            lista = _parse_fragmented_requirements(artefatos_json)
-        except Exception as fallback_e:
-            caminho = _run_async(
-                _gerar_doubt_artifact("ERR_ENTRADA_JSON", f"Erro ao parsear JSON de entrada: {e}")
-            )
-            return {
-                "status": "erro",
-                "mensagem": f"JSON inválido: {e}",
-                "arquivo_duvida": caminho,
-            }
-
-    lista = _normalizar_anexos_inline(lista)
-    lista = _ordenar_por_criticidade(lista)
-    resultados = _run_async(_processar_todos_em_paralelo(lista))
-
-    total     = len(resultados)
-    sucessos  = sum(1 for r in resultados if r["status"] == "sucesso")
-    bloqueados = sum(1 for r in resultados if r["status"] == "bloqueado")
-    falhas    = total - sucessos - bloqueados
-
-    return {
-        "status": "concluido",
-        "resumo": {
-            "total": total,
-            "sucessos": sucessos,
-            "bloqueados": bloqueados,
-            "falhas": falhas,
-        },
-        "detalhes": resultados,
-    }
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# Processamento paralelo
-# ──────────────────────────────────────────────────────────────────────────────
-
-async def _processar_todos_em_paralelo(
-    lista_artefatos: list,
-    max_paralelos: int = 5,
-) -> list:
-    """Processa múltiplos artefatos em paralelo com limite de concorrência.
-
-    Args:
-        lista_artefatos: Lista de artefatos de requisito.
-        max_paralelos: Máximo de processamentos simultâneos (padrão: 5).
-
-    Returns:
-        list: Lista de resultados de cada artefato processado.
-    """
-    semaforo = asyncio.Semaphore(max_paralelos)
-
-    async def processar_com_limite(artefato):
-        async with semaforo:
-            return await _processar_artefato(artefato)
-
-    return await asyncio.gather(*[processar_com_limite(a) for a in lista_artefatos])
-
-
-async def _processar_artefato(artefato: dict) -> dict:
-    """Processa单个 artefato de requisito gerando teste pytest.
-
-    Args:
-        artefato: Dicionário com id_artefato, tipo, conteudo, modulo, criticidade.
-
-    Returns:
-        dict: Resultado com status (sucesso/bloqueado/falha) e caminhos gerados.
-    """
-    id_artefato = artefato.get("id_artefato", "SEM_ID")
-    tipo        = artefato.get("tipo", "RF")
-    conteudo    = artefato.get("conteudo", "")
-    modulo      = artefato.get("modulo", "geral")
-
-    logger.info(f"[QA] Iniciando: {id_artefato} ({tipo})")
-
-    # Valida antes de gerar
-    bloqueio = _validar_artefato(artefato)
-    if bloqueio:
-        caminho = await _gerar_doubt_artifact(id_artefato, bloqueio)
-        logger.warning(f"[QA] Bloqueado: {id_artefato} → {caminho}")
-        return {
-            "id_artefato": id_artefato,
-            "status": "bloqueado",
-            "motivo": "doubt_artifact_gerado",
-            "arquivo_duvida": str(caminho),
-            "mensagem": f"Inconsistência: {bloqueio}. Aguardando intervenção humana.",
-        }
-
-    try:
-        slug = _slugify(id_artefato)
-        artefato_dir = _tests_dir() / slug
-        artefato_dir.mkdir(parents=True, exist_ok=True)
-        (artefato_dir / "__init__.py").touch(exist_ok=True)
-
-        anexos_salvos = _salvar_arquivos_apoio(artefato, artefato_dir)
-        tem_codigo = any(p.suffix in ['.py', '.java', '.js', '.c'] for p in anexos_salvos)
-        
-        nomes_anexos = [p.name for p in anexos_salvos]
-        if nomes_anexos:
-            logger.info(f"[QA] Arquivos anexados detectados para {id_artefato}: {nomes_anexos}")
-        else:
-            logger.info(f"[QA] Nenhum arquivo anexado detectado para {id_artefato}.")
-
-        nome_teste = f"test_{slug}.py"
-        caminho = artefato_dir / nome_teste
-
-        codigo = _gerar_pytest_via_llm(
-            id_artefato=id_artefato,
-            tipo=tipo,
-            conteudo=conteudo,
-            modulo=modulo,
-            arquivos_apoio=anexos_salvos,
-            nome_teste=nome_teste,
-        )
-        codigo_valido = _validar_e_sanitizar_codigo(codigo, id_artefato)
-        caminho.write_text(codigo_valido, encoding="utf-8")
-
-        logger.info(f"[QA] Concluído (Fluxo {'A' if tem_codigo else 'B'}): {id_artefato} → {caminho}")
-        return {
-            "id_artefato": id_artefato,
-            "status": "sucesso",
-            "fluxo": "A" if tem_codigo else "B",
-            "pasta_gerada": str(artefato_dir),
-            "arquivo_gerado": str(caminho),
-            "arquivos_apoio": [str(p) for p in anexos_salvos],
-            "erro": None,
-        }
-
-    except Exception as e:
-        logger.error(f"[QA] Erro em {id_artefato}: {e}")
-        return {
-            "id_artefato": id_artefato,
-            "status": "falha",
-            "arquivo_gerado": None,
-            "erro": str(e),
-        }
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# Validação
-# ──────────────────────────────────────────────────────────────────────────────
-
-def _validar_artefato(artefato: dict) -> str | None:
-    """Valida artefato de requisito verificando campos obrigatórios.
-
-    Args:
-        artefato: Dicionário com campos do artefato.
-
-    Returns:
-        str | None: Descrição do bloqueio ou None se válido.
-    """
-    if not artefato.get("conteudo", "").strip():
-        return "Campo 'conteudo' vazio — impossível gerar testes sem descrição do requisito."
-    if not artefato.get("modulo", "").strip():
-        artefato["modulo"] = "geral"
-    if artefato.get("tipo") not in ("RF", "HU", "UC", "RNF", "RN"):
-        return (
-            f"Tipo desconhecido: '{artefato.get('tipo')}'. "
-            "Esperado: RF, HU, UC, RNF ou RN."
-        )
-    return None
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# Doubt Artifact
-# ──────────────────────────────────────────────────────────────────────────────
-
-async def _gerar_doubt_artifact(id_artefato: str, motivo: str) -> str:
-    """Gera arquivo de doubt artifact para artefato bloqueado.
-
-    Args:
-        id_artefato: Identificador do artefato.
-        motivo: Motivo do bloqueio.
-
-    Returns:
-        str: Caminho do arquivo gerado.
-    """
-    timestamp = datetime.now(timezone.utc).strftime("%Y%m%dT%H%M%S")
-    _doubt_dir().mkdir(parents=True, exist_ok=True)
-
-    nome = f"Doubt_Artifact_{id_artefato}_{timestamp}.md"
-    caminho = _doubt_dir() / nome
-
-    conteudo = f"""# Doubt Artifact — QA Agent
-
-**ID do Artefato:** {id_artefato}
-**Data/Hora:** {timestamp}
-**Agente:** qa_agent
-**Status:** BLOQUEADO — aguardando intervenção humana
-
----
-
-## Descrição do Bloqueio
-
-{motivo}
-
-## O que é necessário para continuar
-
-[ Preencher após intervenção ]
-
-## Resolução
-
-- **Resolvido por:** [ Preencher ]
-- **Data:** [ Preencher ]
-- **Ação tomada:** [ Preencher ]
-"""
-    caminho.write_text(conteudo, encoding="utf-8")
-    return str(caminho)
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# Geração do template pytest
-# ──────────────────────────────────────────────────────────────────────────────
-
-def _slugify(texto: str) -> str:
-    """Normaliza texto para uso em nomes de arquivo.
-
-    Args:
-        texto: Texto a normalizar.
-
-    Returns:
-        str: Slug seguro para nomes de arquivo.
-    """
-    base = (texto or "artefato").strip().lower()
-    base = re.sub(r"[^a-z0-9]+", "_", base)
-    base = base.strip("_")
-    return base or "artefato"
-
-
-def _safe_filename(nome: str) -> str:
-    """Sanitiza nome de arquivo removendo caracteres especiais.
-
-    Args:
-        nome: Nome original do arquivo.
-
-    Returns:
-        str: Nome seguro para filesystem.
-    """
-    cleaned = re.sub(r"[^a-zA-Z0-9._-]", "_", (nome or "arquivo.txt").strip())
-    cleaned = cleaned.lstrip(".")
-    return cleaned or "arquivo.txt"
-
-
-def _salvar_arquivos_apoio(artefato: dict, destino: Path) -> list[Path]:
-    """Salva arquivos de apoio (texto ou base64) no diretório de destino.
-
-    Args:
-        artefato: Dicionário contendo lista de arquivos_apoio.
-        destino: Path do diretório onde arquivos serão salvos.
-
-    Returns:
-        list[Path]: Lista de paths dos arquivos salvos.
-    """
-    arquivos = artefato.get("arquivos_apoio", [])
-    if not isinstance(arquivos, list):
-        return []
-
-    salvos: list[Path] = []
-    for item in arquivos:
-        if not isinstance(item, dict):
-            continue
-
-        nome = _safe_filename(item.get("nome") or item.get("filename") or "arquivo.txt")
-        conteudo_texto = item.get("conteudo")
-        conteudo_b64 = item.get("conteudo_base64")
-
-        caminho = destino / nome
-
-        if isinstance(conteudo_texto, str):
-            caminho.write_text(conteudo_texto, encoding="utf-8")
-            salvos.append(caminho)
-            continue
-
-        if isinstance(conteudo_b64, str):
-            try:
-                bruto = base64.b64decode(conteudo_b64)
-            except Exception:
-                continue
-            caminho.write_bytes(bruto)
-            salvos.append(caminho)
-
-    return salvos
-
-
-def _gerar_pytest_via_llm(
-    id_artefato: str,
-    tipo: str,
-    conteudo: str,
-    modulo: str,
-    arquivos_apoio: list[Path],
-    nome_teste: str,
-) -> str:
-    """Gera código pytest usando LLM a partir de artefato de requisito.
-
-    Args:
-        id_artefato: Identificador do artefato.
-        tipo: Tipo do artefato (RF, HU, UC, RNF, RN).
-        conteudo: Conteúdo do requisito.
-        modulo: Módulo alvo do teste.
-        arquivos_apoio: Lista de paths de arquivos de apoio.
-        nome_teste: Nome do arquivo de teste a gerar.
-
-    Returns:
-        str: Código Python do teste pytest.
-
-    Raises:
-        ValueError: Se o modelo retornar conteúdo vazio.
-    """
-    model_name = os.environ.get("ADK_LLM_MODEL", "gemini-2.5-flash")
-    llm_kwargs = {}
-    if "/" not in model_name:
-        model_name = f"gemini/{model_name}"
-        llm_kwargs["api_key"] = os.environ.get("GOOGLE_API_KEY")
-    arquivos_textos = []
-    for p in arquivos_apoio:
-        try:
-            texto = p.read_text(encoding="utf-8")
-            arquivos_textos.append(f"--- {p.name} ---\n{texto}\n")
-        except Exception:
-            arquivos_textos.append(f"- {p.name} (Arquivo binário ou ilegível)")
-
-    arquivos_desc = "\n".join(arquivos_textos)
-    contexto_arquivos = (
-        "Arquivos de apoio e CÓDIGO FONTE fornecidos para o teste:\n"
-        f"{arquivos_desc}\n"
-        if arquivos_desc
-        else "Nenhum arquivo de apoio ou código fonte foi fornecido.\n"
-    )
-    tem_codigo = any(p.suffix in ['.py', '.java', '.js', '.c'] for p in arquivos_apoio)
-    
-    if tem_codigo:
-        instrucao_geracao = (
-            "O usuário forneceu o código fonte junto aos requisitos. "
-            "MAPEAMENTO: Mapeie os cenários de teste contra as funções e métodos reais presentes no código. "
-            "Gere os testes pytest COMPLETOS e integrados, utilizando asserts que validem as lógicas existentes. "
-            "REGRA DE IMPORTAÇÃO MANDATÓRIA: Faça a importação das funções/classes de forma RELATIVA e EXPLÍCITA a partir do arquivo fornecido. "
-            "Exemplo obrigatório: se o arquivo for 'calculadora.py' com a função 'somar', você DEVE usar: `from .calculadora import somar`."
-        )
-    else:
-        instrucao_geracao = (
-            "Nenhum código fonte foi fornecido — gere em MODO ESQUELETO. "
-            "Use @pytest.mark.skip(reason='Aguardando implementação do código fonte') "
-            "antes de cada função de teste. O corpo deve ter apenas uma docstring "
-            "descrevendo o comportamento esperado. NÃO use 'pass' — a docstring é "
-            "o corpo válido da função em Python."
-        )
-
-    prompt = f"""Gere SOMENTE código Python válido para {nome_teste}.
-Artefato: {id_artefato}
-Tipo: {tipo}
-Módulo alvo: {modulo}
-Requisito: {conteudo}
-
-{contexto_arquivos}
-
-DIRETRIZ DE GERAÇÃO CONDICIONAL: 
-{instrucao_geracao}
-
-Regras obrigatórias:
-- Retorne apenas código Python, sem markdown.
-- Use pytest.
-- O teste deve ser executável mesmo sem instalação de módulos externos ao diretório local.
-- Se houver arquivo-fonte local, faça import relativo via pathlib/sys.path usando a própria pasta do teste.
-- Se não houver código-fonte importável, gere testes de contrato (validações e comportamentos inferíveis) sem import quebrado.
-- Cubra cenários feliz, inválido e borda.
-- Inclua asserts objetivos.
-- Cada função de teste deve ter corpo NÃO-VAZIO: ou uma docstring (modo
-  esqueleto), ou asserts objetivos (modo completo). Nunca emita 'pass'
-  isolado, 'TODO', placeholders entre <>, ou caracteres fora da gramática Python.
-"""
-
-    response = completion(
-        model=model_name,
-        messages=[
-            {
-                "role": "system",
-                "content": "Você gera exclusivamente código de teste pytest executável.",
-            },
-            {
-                "role": "user",
-                "content": prompt,
-            },
-        ],
-        temperature=0,
-        **llm_kwargs,
-    )
-
-    codigo = ""
-    choices = getattr(response, "choices", None)
-    if choices and len(choices) > 0:
-        message = getattr(choices[0], "message", None)
-        codigo = getattr(message, "content", "") if message else ""
-
-    if not codigo and isinstance(response, dict):
-        response_choices = response.get("choices", [])
-        if response_choices:
-            codigo = response_choices[0].get("message", {}).get("content", "")
-
-    if not isinstance(codigo, str) or not codigo.strip():
-        raise ValueError("Modelo retornou conteúdo vazio para geração de pytest.")
-
-    return codigo.strip() + "\n"
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# Priorização
-# ──────────────────────────────────────────────────────────────────────────────
-
-def _ordenar_por_criticidade(lista: list) -> list:
-    """Ordena lista de artefatos por criticidade (alta > media > baixa).
-
-    Args:
-        lista: Lista de artefatos com campo criticidade.
-
-    Returns:
-        list: Lista ordenada por prioridade.
-    """
-    prioridade = {"alta": 0, "media": 1, "baixa": 2}
-    return sorted(lista, key=lambda a: prioridade.get(a.get("criticidade", "media"), 1))
-
-
-agent = LlmAgent(
-    name="receber_requisitos",
-    model=os.environ.get("ADK_LLM_MODEL", "gemini-2.5-flash"),
-    description=(
-        "Subagente que recebe artefatos de requisito em JSON e gera arquivos pytest "
-        "funcionais em workspace_output/tests/inputs/."
-    ),
-    instruction=(
-    "Ao receber uma mensagem do usuário, monte um JSON com os campos: "
-    "id_artefato (ex: 'HU-001'), tipo ('HU'), conteudo, modulo ('geral' se não informado), criticidade ('alta'). "
-    "MUITO IMPORTANTE: Se a requisição contiver código-fonte (em anexo ou no texto), você DEVE criar uma propriedade chamada 'arquivos_apoio' "
-    "sendo uma lista de objetos com 'nome' (ex: arquivo.py) e 'conteudo' (o código completo). "
-    "Se você não preencher 'arquivos_apoio', o sistema NÃO reconhecerá o código-fonte! "
-    "No campo conteudo, inclua apenas o texto do requisito. "
-    "Chame a tool receber_requisitos com o JSON gerado e retorne o resultado."
-),
-    tools=[
-        FunctionTool(receber_requisitos),
-    ],
-)
